fix update tag and timestamp tracing
For simplicity update_tag and timestamp are now stored as uint64_t, to
avoid corner cases.

Furthermore, they are also reported to the trace as uint64_t (instead of
double as before). The old behavior produced rounding errors and should
have not been used in the first place.

+++ include/occ_util.hpp
@@ -68,7 +68,7 @@ struct sensor_data_t {
     uint64_t accumulator;
 
     /// number of samples stored in accumulator
-    uint32_t update_tag;
+    uint64_t update_tag;
 };
 using sensor_data_t = struct sensor_data_t;
